refactor: distinct를 tag로 변경
