+++ src/components/Main/Main.jsx
@@ -4,7 +4,7 @@ import styles from './Main.module.css';
 import useClass from './hooks/useClass';
 import BottomSheet from '../common/BottomSheet/BottomSheet';
 import Body from '../common/BottomSheet/Body/Body';
-import { ATTEND, CREATE } from '../../constants/bottomSheetDistinct';
+import { ATTEND, CREATE } from '../../constants/bottomSheetTag';
 
 export default function Main() {
   const {
@@ -23,12 +23,12 @@ export default function Main() {
 
       <BottomSheet>
         <BottomSheet.Header>
-          <BottomSheet.Button text='모임 만들기' type='button' distinct={CREATE} />
-          <BottomSheet.Button text='모임 참여하기' type='button' distinct={ATTEND} />
+          <BottomSheet.Button text='모임 만들기' type='button' tag={CREATE} />
+          <BottomSheet.Button text='모임 참여하기' type='button' tag={ATTEND} />
         </BottomSheet.Header>
         <BottomSheet.Body>
-          <Body.ClassCreate distinct={CREATE} />
-          <Body.ClassAttend distinct={ATTEND} />
+          <Body.ClassCreate tag={CREATE} />
+          <Body.ClassAttend tag={ATTEND} />
         </BottomSheet.Body>
       </BottomSheet>
     </>
